refactor: move generate_kfold_splits to utils.data_utils
